2022.9.23 第三次提交  => 合并分支 修订部分实体id为Long  => 第一次测试配置

+++ deloitte-modules/inner-crm/src/main/java/com/deloitte/crm/domain/CrmWindDict.java
@@ -19,7 +19,7 @@ public class CrmWindDict extends BaseEntity
     private static final long serialVersionUID = 1L;
 
     /** $column.columnComment */
-    private Long id;
+    private Integer id;
 
     /** wind文件分类名 */
     @Excel(name = "wind文件分类名")
@@ -56,12 +56,12 @@ public class CrmWindDict extends BaseEntity
     @Excel(name = "${comment}", readConverterExp = "$column.readConverterExp()")
     private Date updated;
 
-    public void setId(Long id) 
+    public void setId(Integer id)
     {
         this.id = id;
     }
 
-    public Long getId() 
+    public Integer getId()
     {
         return id;
     }
